+++ src/main/java/epicode/u5d01introductionspringboot/model/toppings/ToppingDecorator.java
@@ -20,7 +20,7 @@ public class ToppingDecorator extends Consummation {
     return this.decoratedPizza.getPrice() + this.price;
   }
 
-  public long getCalories() {
+  public double getCalories() {
     return this.decoratedPizza.getCalories() + this.calories;
   }
